Moved tests for moved util functions
